Use bytes interface value for empyt native type struct fields instead of dynamic list
If a struct has a byte array/slice field and it happens to be the zero
value, a dynamic list object was returned. This caused issues with
functions taking `Bytes` via arguments, as the type checker will accept
such functions, but they might fail at runtime.

To work around this, the raw bytes array/slice is returned in case of an
zero value.